+++ test/integration/sensor.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2014-2015 Open Source Robotics Foundation
+ * Copyright (C) 2014-2016 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -33,6 +33,19 @@ TEST_F(SensorTest, GetScopedName)
 
   std::string sensorName = sensor->ScopedName();
   EXPECT_EQ(sensorName, std::string("default::rotated_box::link::cam1"));
+}
+
+/////////////////////////////////////////////////
+// Make sure sensors can run without asserting in a world with a large
+// step size.
+TEST_F(SensorTest, FastSensor)
+{
+  Load("worlds/fast_sensor_test.world");
+  physics::WorldPtr world = physics::get_world("default");
+  ASSERT_TRUE(world != NULL);
+
+  // This test will cause an assertion if maxSensorUpdate in
+  // SensorManager::SensorContainer::RunLoop() is set improperly
 }
 
 int main(int argc, char **argv)