+++ Changelog.md
@@ -1,4 +1,15 @@
-## Gazebo 7.0
+## Gazebo 7
+
+## Gazebo 7.x.x (2016-xx-xx)
+
+1. Fix SensorManager::SensorContainer::RunLoop sensor update time assertion
+    * [Pull request #2115](https://bitbucket.org/osrf/gazebo/pull-request/2115)
+
+1. Removes one function call and replaces a manual swap with std::swap in ODE heightfield.
+    * [Pull request #2114](https://bitbucket.org/osrf/gazebo/pull-request/2114)
+
+
+## Gazebo 7.0.0 (2016-01-25)
 
 1. Add FollowerPlugin
     * [Pull request #2085](https://bitbucket.org/osrf/gazebo/pull-request/2085)
@@ -342,7 +353,7 @@ using the same arguments used in the command line.
 
 ### Gazebo 6.X.X (201X-XX-XX)
 
-1. Fix removing multiple camera sensors that have the same camera name
+1. Fix removing multiple camera sensors that have the same camera name.
     * [Pull request #2081](https://bitbucket.org/osrf/gazebo/pull-request/2081)
     * [Issue #1811](https://bitbucket.org/osrf/gazebo/issues/1811)